% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Make_Decoys.R
\name{allDecoys}
\alias{allDecoys}
\title{Create decoy database}
\usage{
allDecoys(Identified, version, max.tails, exact.tails, ion.mode = "Pos")
}
\arguments{
\item{Identified}{Identified lipids file}

\item{version}{Appends this character to the output file name of identified lipids.}

\item{max.tails}{list of desired number of carbons and the maximum number of double
bonds for each fatty acyl chain length.}

\item{exact.tails}{library can be customizable for specific tails}

\item{ion.mode}{Pos or Neg to indicate positive or negative ion mode}
}
\value{
a library of decoy triacylglycerols
}
\description{
Create decoy database
}
\examples{
\dontrun{
Decoys <- allDecoys(version = '2024',
 ion.mode = 'Pos', max.tails = "8.2, 9.0, 10.2, 11.0, 12.3, 13.1, 14.3, 15.3,
  16.5, 17.3, 18.5, 19.5, 20.6, 21.5, 22.6, 23.0, 24.4, 25.0, 26.0", 
  Identified = Lipids)
  }
}
