% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_FDRCutoff.R
\name{FDRCutoff}
\alias{FDRCutoff}
\title{Generating a list of TGs passed through a pre-determined FDR}
\usage{
FDRCutoff(identified.lipids, standard.file = NULL, iteration = 3, cutoff = 0.1)
}
\arguments{
\item{identified.lipids}{List of combined TG annotations}

\item{standard.file}{list of standard TGs (Optional)}

\item{iteration}{#Default 10}

\item{cutoff}{Empirical FDR. Default 0.1}
}
\value{
list of TGs that pass through the FDR cutoff
}
\description{
Generating a list of TGs passed through a pre-determined FDR
}
\examples{
 \dontrun{
ResultsTG_FDR <- FDRCutoff(identified.lipids=ResultsTG,
standard.file = "Standards.csv",
iteration=10,
cutoff = 0.1)
}
}
