% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_LipidIdentifier.R
\name{LipidIdentifier}
\alias{LipidIdentifier}
\title{TG identification}
\usage{
LipidIdentifier(
  DIADataObj,
  lipid,
  ion.mode,
  format = c("MSDIAL", "Generic", "Progenesis"),
  ppmtol = 15,
  ppmtolMS1,
  rttol = 5,
  intensity.window = c(0, Inf),
  which.frag = "any",
  ms1.precursors = c(1),
  ms2.precursors = c(0, 1),
  version,
  max.tails =
    "8.2, 9.0, 10.2, 11.0, 12.3, 13.1,14.3, 15.3, 16.5, 17.3, 18.5, 19.5, 20.6, 21.5, 22.6, 23.0, 24.4, 25.0, 26.0",
  exact.tails = NULL,
  print.MS2spectra = FALSE,
  write.annotations = TRUE
)
}
\arguments{
\item{DIADataObj}{The object stores alignment file and spectra}

\item{lipid}{= "TAG" #DIATAGeR is for triacylglycerols only.}

\item{ion.mode}{Pos or Neg to indicate positive or negative ion mode}

\item{format}{tell the function if the alignment file is MSDIAL or Generic Format}

\item{ppmtol}{ppm tolerance. Default to 15}

\item{ppmtolMS1}{Defaults to ppmtol if not specified.}

\item{rttol}{retention time tolerance in seconds. Default to 5.}

\item{intensity.window}{Two numerical values, specifying the percentage multiplication
factor for the precursor intensity that the fragments must fall within. If set
to c(0, 100), the fragment intensity cannot be greater than the precursor intensity.
Defaults to 0 and infinity.}

\item{which.frag}{Either 'any' or 'firsttwo', specifying whether any fragment
should be used or if the first two fragments should be used for identification.
Useful when there are three possible fragments in the reference and only two are
required for a positive match. Default 'any'.}

\item{ms1.precursors}{Number of precursors found in MS1 to identify compounds. Default to 1}

\item{ms2.precursors}{Number of precursor found in MS2 to identify compounds.
"0" is suggested for high collision energy and "1" for low collision energy.}

\item{version}{Appends this character to the output file name.}

\item{max.tails}{list of desired number of carbons and the maximum number of double
bonds for each fatty acyl chain length.
Default = "8.2, 9.0, 10.2, 11.0, 12.3, 13.1, 14.3, 15.3, 16.5, 17.3,
18.5, 19.5, 20.6, 21.5, 22.6, 23.0, 24.4, 25.0, 26.0"}

\item{exact.tails}{library can be customizable for specific tails. (Eg: exact_tails = c("18.1","18.2) creates library of TGs made up 18:1 and 18:2 fatty acids))}

\item{print.MS2spectra}{If TRUE, prints MS/MS spectra and mirrored reference peaks. Default FALSE.}

\item{write.annotations}{If TRUE, writes RDS file of identified lipids. Default TRUE.}
}
\value{
A folder named "Pos" or "Neg" (depends on the ion mode) will be created and contain a list of identified triacylglycerols
}
\description{
TG identification
}
\examples{

\dontrun{
LipidIdentifier(DIADataObj = DIA_Pos,
lipid = "TAG", 
ion.mode = "Pos", 
format = "MSDIAL",
ppmtol=15, rttol=5, 
intensity.window = c(0,Inf),
version = "24", 
ms1.precursors= 1, ms2.precursors = 1,
max.tails = "8.2, 9.0, 10.2, 11.0, 12.3, 13.1, 14.3, 15.3, 16.5, 17.3, 
18.5, 19.5, 20.6, 21.5, 22.6, 23.0, 24.4, 25.0, 26.0",
exact.tails=NULL,
which.frag = "any", 
print.MS2spectra = F,
write.annotations = T)
}
}
