% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_measureFileImport.R
\name{measureFileImport}
\alias{measureFileImport}
\title{Import the feature list (MSDIAL alignment file or Generic Format)}
\usage{
measureFileImport(
  filename,
  ResultsFiletype = c("MSDIAL", "Generic"),
  IonMode = c("Pos", "Neg"),
  AlignmentFileName = NULL
)
}
\arguments{
\item{filename}{MSDIAL alignment file or Generic Import Format (.txt file)}

\item{ResultsFiletype}{tell the function if the import file is MSDIAL Format or Generic Format}

\item{IonMode}{Pos or Neg to indicate positive or negative ion mode}

\item{AlignmentFileName}{NULL}
}
\value{
Imported feature list
}
\description{
Import the feature list (MSDIAL alignment file or Generic Format)
}
\examples{
\dontrun{
DIA_Pos <-measureFileImport(filename = "Height_0_20239151316.txt",
IonMode = "Pos", ResultsFiletype = "MSDIAL")
}
}
