% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_RemoveIsotopes.R
\name{RemoveIsotopes}
\alias{RemoveIsotopes}
\title{Deisotoping and denosing MS2 spectra}
\usage{
RemoveIsotopes(
  DIADataObj,
  ppmtol = 15,
  low.intensity.cutoff = 100,
  density.cutoff = T,
  density.cutoff.value = 1,
  check.isotopes = T,
  lenient.removal = T
)
}
\arguments{
\item{DIADataObj}{object stores feature list and MS2 spectra.}

\item{ppmtol}{ppm tolerance. Default 15.}

\item{low.intensity.cutoff}{Anything below this value will be removed. Default 100.}

\item{density.cutoff}{If TRUE, a density-based cutoff will be applied. The kernel
density estimate of each spectra's intensity distribution will be used to implement
a low intensity cutoff.}

\item{density.cutoff.value}{The low intensity cutoff is the intensity where the
percentage change in the the Riemann sum of the kernel density when the intensity
is increased by 10 drops below the density.cutoff.value. Higher value for a lower intensity
cutoff. Default 1. Change with caution.}

\item{check.isotopes}{If TRUE, check the isotopes in the spectra. Default TRUE}

\item{lenient.removal}{If TRUE, removes M+1 and M+2 isotope peaks. Default TRUE}
}
\value{
De-isotoped spectra
}
\description{
Deisotoping and denosing MS2 spectra
}
\examples{
\dontrun{
DIA_Pos <- KeepIsotopes(DIADataObj = DIA_Pos, 
ppmtol = 15,
low.intensity.cutoff=100,
density.cutoff=T,
density.cutoff.value=1)
}
}
