% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_ScoreLipids.R
\name{ScoreLipids}
\alias{ScoreLipids}
\title{Generate Scores for ranking lipids.}
\usage{
ScoreLipids(
  DIADataObj,
  ion.mode,
  format = c("MSDIAL", "Generic", "Progenesis"),
  lipid,
  version,
  ppmtol = 15,
  ppmtolMS1,
  rttol = 5,
  run.code = "Sample",
  ms1.precursors = c(1),
  ms2.precursors = c(0, 1),
  spectra.file.type = c("txt", "msp", "mgf", "mgf_mzmine"),
  max.tails =
    "8.2, 9.0, 10.2, 11.0, 12.3, 13.1,14.3, 15.3, 16.5, 17.3, 18.5, 19.5, 20.6, 21.5, 22.6, 23.0, 24.4, 25.0, 26.0",
  exact.tails = NULL,
  spectra.file,
  ...
)
}
\arguments{
\item{DIADataObj}{The object stores alignment file and spectra}

\item{ion.mode}{Mass spectrometry ion mode, the same as in LipidIdentifier.}

\item{format}{tell the function if the alignment file is MSDIAL or Generic Format}

\item{lipid}{Lipid class, the same as in LipidIdentifier.}

\item{version}{Version name, the same as in LipidIdentifier. Also used to pull identified lipids from the ion.mode folder.}

\item{ppmtol}{PPM tolerance, the same as in LipidIdentifier. Default 15.}

\item{ppmtolMS1}{Defaults to ppmtol if not specified.}

\item{rttol}{Retention time tolerance, the same as in LipidIdentifier. Default 5.}

\item{run.code}{for calculating the correlation between fragments and precursor intensities.
This helps recognize the columns of samples in the alignment file. Default "Sample"}

\item{ms1.precursors}{Number of precursors found in MS1 to identify compounds. Default to 1}

\item{ms2.precursors}{Number of precursor found in MS2 to identify compounds.}

\item{spectra.file.type}{specify txt, msp or mgf files. The user will have determined the spectra name}

\item{max.tails}{list of desired number of carbons and the maximum number of double
bonds for each fatty acyl chain length.
Default = "8.2, 9.0, 10.2, 11.0, 12.3, 13.1, 14.3, 15.3, 16.5, 17.3, 18.5,
19.5, 20.6, 21.5, 22.6, 23.0, 24.4, 25.0, 26.0"}

\item{exact.tails}{library can be customizable for specific tails}

\item{spectra.file}{File path for retrieving spectra from all samples.}

\item{...}{Additional parameters}
}
\value{
scores for targeted and decoy triacylglycerols
}
\description{
Generate Scores for ranking lipids.
}
\examples{
\dontrun{
ResultsTAG <- ScoreLipids(DIADataObj = DIA_Pos,
lipid = "TAG", 
ion.mode = "Pos",
version = "24", 
ppmtol= 15, 
rttol = 5, 
ms1.precursors =1, ms2.precursors =1, 
spectra.file.type ="txt",
max.tails = "8.2, 9.0, 10.2, 11.0, 12.3, 13.1, 14.3, 15.3, 16.5, 17.3, 
18.5, 19.5, 20.6, 21.5, 22.6, 23.0, 24.4, 25.0, 26.0",
spectra.file = "Data/Centroid/")
}

}
