% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_SpectraImport.R
\name{SpectraImport}
\alias{SpectraImport}
\title{MS2 spectra import}
\usage{
SpectraImport(
  fileORfoldername,
  DIADataObj,
  ion.mode,
  results.file.type = c("MSDIAL", "Generic", "Progenesis"),
  spectra.type = "centroid",
  spectra.file.type = c("txt", "msp", "mgf"),
  ppmtol = 5,
  rttol = 2
)
}
\arguments{
\item{fileORfoldername}{path to the folder that stores centroid MS2 spectra}

\item{DIADataObj}{object stores the feature list}

\item{ion.mode}{Pos or Neg to indicate positive or negative ion mode}

\item{results.file.type}{tell the function if the feature list is MSDIAL alignment file or Generic Format}

\item{spectra.type}{Only for centroid data. The user will have determined the spectra type
when exporting from MSDIAL so this is more of a reminder and will ensure proper functions are called.}

\item{spectra.file.type}{Specify txt, msp or mgf files. The user will have determined the spectra name
when exporting from MSDIAL so this is more of a reminder and will ensure proper functions are called.}

\item{ppmtol}{ppm tolerance. Default to 5. This parameter is for for generic format since it retrieve the spectra based on rt and m/z similarity}

\item{rttol}{retention time tolerance in seconds.}
}
\value{
MSMS spectra of feature list
}
\description{
MS2 spectra import
}
\examples{
\dontrun{
DIA_Pos <- SpectraImport(DIADataObj = DIA_Pos, 
fileORfoldername =  "Data/centroid",
ion.mode = "Pos", 
spectra.file.type ="txt",
results.file.type = "MSDIAL",
rttol = 10)
}
}
