% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_makeTAGs.R
\name{getTAGs}
\alias{getTAGs}
\title{Create triacylglycerols library}
\usage{
getTAGs(
  tails =
    "8.2, 9.0, 10.2, 11.0, 12.3, 13.1,14.3, 15.3, 16.5, 17.3, 18.5, 19.5, 20.6, 21.5, 22.6, 23.0, 24.4, 25.0, 26.0",
  exact_tails = NULL
)
}
\arguments{
\item{tails}{list of desired number of carbons and the maximum number of double
bonds for each fatty acyl chain length. Format: number of carbons. number of
double bonds (Eg: tails = 8.2 creates three fatty acyl chains: 8:0, 8:1 and 8:2)}

\item{exact_tails}{library can be customizable for specific tails
(Eg: exact_tails = c("18.1","18.2) creates library of TGs made up 18:1 and 18:2 fatty acids)}
}
\value{
library of triacylglycerols with information of precursors and fragments masses
}
\description{
Create triacylglycerols library
}
\examples{
\dontrun{
Library <- getTAGs(tails = "8.2, 9.0, 10.2, 11.0, 12.3, 13.1, 14.3, 15.3, 
16.5, 17.3, 18.5, 19.5, 20.6, 21.5, 22.6, 23.0, 24.4, 25.0, 26.0")
}

}
